STY: rename class
