+++ pysat/tests/classes/cls_registration.py
