+++ pysat/tests/test_constellation.py
@@ -12,7 +12,7 @@ import pytest
 
 import pysat
 from pysat import constellations
-from pysat.tests.classes.cls_registration_tests import TestWithRegistration
+from pysat.tests.classes.cls_registration import TestWithRegistration
 
 
 class TestConstellationInitReg(TestWithRegistration):
