+++ pysat/tests/test_utils.py
@@ -16,7 +16,7 @@ import shutil
 import tempfile
 
 import pysat
-from pysat.tests.classes.cls_registration_tests import TestWithRegistration
+from pysat.tests.classes.cls_registration import TestWithRegistration
 from pysat.utils import generate_instrument_list
